Folding addAuthenticationToHttpClient in as a static method.
